Refactor awaitAndReceive
Avoid deserializing first to `String` and then to the desirable `T`,
since it triggers to deserializations.

Note that, since we use `kotlinx.serialization` for serialization, the
library is also decoding the request in two steps, but this is a limit
from the library's side and probably will be lifted in the future.

Signed-off-by: George Papadopoulos <george.719pap@gmail.com>